+++ Tp3_omar_alex/src/GenerateurNombreAleatoire.hpp
@@ -3,6 +3,7 @@
 
 #include <iostream>
 #include "Dvector.h"
+#include "stdint.h"
 
 class GenerateurNombreAleatoire{
     private:
@@ -11,13 +12,13 @@ class GenerateurNombreAleatoire{
 
     public:
     GenerateurNombreAleatoire();
-    GenerateurNombreAleatoire(int dim, int rand_max);
+    GenerateurNombreAleatoire(int dim, uint64_t rand_max);
     GenerateurNombreAleatoire(const GenerateurNombreAleatoire & toCopy);
     ~GenerateurNombreAleatoire();
     int get_dim();
     void set_dim(int dim);
     GenerateurNombreAleatoire & clone(const GenerateurNombreAleatoire & toCopy);
-    int get_rand_max();
+    uint64_t get_rand_max();
     virtual Dvector generate() = 0;
     virtual void set_seed(int seed) = 0;
     virtual int get_seed() = 0;
