+++ Tp3_omar_alex/src/XorShift.hpp
@@ -1,11 +1,12 @@
 #ifndef XORSHIFT_HPP
 #define XORSHIFT_HPP
 
+#include "stdint.h"
 #include "GenerateurNombreAleatoire.hpp"
 
 class XorShift{
     private:
-    int seed;
+    uint64_t seed;
     int seed0; //le seed initial, pour pouvoir revenir avec reset_seed
     
     public:
@@ -14,7 +15,7 @@ class XorShift{
     XorShift(const XorShift & toCopy);
     // ~XorShift();
 
-    int generate_int();
+    uint64_t generate_int();
     void set_seed(int seed);
     int get_seed() const;
     void reset_seed();
